requirements.txt
django
channels
channels_redis
gunicorn
daphne
mysqlclient
pillow
geocoder
django-ipware